Create Fetching_API_Data
Fetching data from Github's API server and isolating the data that is relative to the target of the project.

Target: To analyze/visualize the top 30 highest rated (star system) projects in python that are currently hosted on Github. Naturally, this data will change over time since live data is being fetched every time the program is ran.